refactor: 발행했는지 확인 로직 PublishService클래스로 이동
